chore: move train to extension
